requirements.txt
pandas
openpyxl
streamlit
altair<5
fpdf2